requirements.txt
pytest
pylint
python-dotenv
boto3
pandas
psycopg2
ipykernel